Move TP array context to array_context.py, other minor changes
